# docs/swagger.yaml
definitions:
...
  models.Comment:
    properties:
      comment:
        description: Comment comment
        type: string
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      id:
        description: Model id pk
        type: integer
      post:
        $ref: '#/definitions/models.Post'
      post_id:
        description: Comment postid
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.CommentForCreate:
    properties:
      comment:
        description: Comment comment
        type: string
      post_id:
        description: Comment postid
        type: string
    type: object
  models.CommentForUpdate:
    properties:
      comment:
        description: Comment comment
        type: string
      post_id:
        description: Comment postid
        type: string
    type: object
...
  models.Permission:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      desc:
        description: Permission desc
        example: คำอธิบาย
        type: string
      id:
        description: Permission id
        example: 1
        type: integer
      name:
        description: Permission name
        example: create table
        type: string
      slug:
        description: Permission slug
        example: /aa
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.Post:
    properties:
      comments:
        description: HasMany
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      desc:
        description: Post desc
        type: string
      id:
        description: Model id pk
        type: integer
      title:
        description: Post title
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: Post useri
        type: string
    type: object
...
  models.Posts:
    properties:
      posts:
        items:
          $ref: '#/definitions/models.Posts'
        type: array
    type: object
  models.Role:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      id:
        description: Role id
        example: 1
        type: integer
      name:
        description: Role name
        example: Admin
        type: string
      permissions:
        description: Role permission
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.User:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      email:
        description: User email
        example: a@email.com
        type: string
      id:
        description: Model id pk
        type: integer
      level:
        description: User level
        example: 99
        type: integer
      name:
        description: User password
        example: 'Mr. A '
        type: string
      password:
        description: User password
        example: P@ssw0rd999
        type: string
      posts:
        description: user hasMany  Post  เอา  user_id
        items:
          $ref: '#/definitions/models.Posts'
        type: array
      reftoken:
        description: User token
        example: aaaaa.bbbbbb.cccccc
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      telephone:
        description: User telephone
        example: "0816477729"
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
      username:
        description: User name
        example: usera
        type: string
    type: object
...
  utils.ResponseData:
    properties:
      status:
        description: Response status
        type: integer
      success:
        default: true
        description: Response success
        example: true
        type: boolean
    type: object
info:
  contact:
    email: limweb@hotmail.com
    name: API Support
    url: http://www.servit.co.th/support
  description: '# Manageing Customer API'
  license:
    name: The MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://www.servit.co.th/
  title: ApiTest
  version: "1.0"
paths:
...
  /api/v1/comments/all:
    get:
      consumes:
      - application/json
      description: List all comments
      operationId: ListComments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Comment'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Comments
      tags:
      - comments
  /api/v1/comments/by/{id}:
    get:
      consumes:
      - application/json
      description: Get comment by id
      operationId: GetComment
      parameters:
      - description: id of comment to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Basic auth: []
      summary: Get Comment
      tags:
      - comments
  /api/v1/comments/del/{id}:
    get:
      consumes:
      - application/json
      description: Delete comment by id
      operationId: DeleteComment
      parameters:
      - description: id of comment to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - comments
  /api/v1/comments/edit/{id}:
    post:
      consumes:
      - application/json
      description: Update comment by id
      operationId: UpdateComment
      parameters:
      - description: id of comment to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Comment data to be updated
        in: body
        name: Comment
        required: true
        schema:
          $ref: '#/definitions/models.CommentForUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update Comment
      tags:
      - comments
  /api/v1/comments/last:
    get:
      consumes:
      - application/json
      description: List all comments
      operationId: ListlastComments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Comment'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Comments
      tags:
      - comments
  /api/v1/comments/new:
    post:
      consumes:
      - application/json
      description: Create new comment
      operationId: CreateComment
      parameters:
      - description: Comment data to be created
        in: body
        name: Comment
        required: true
        schema:
          $ref: '#/definitions/models.CommentForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - comments
  /api/v1/comments/vuetable:
    get:
      consumes:
      - application/json
      description: List all comments with pagination keyword filter by column Sort
        columns
      operationId: vuetableComment
      parameters:
      - description: number of page
        in: query
        name: page
        type: string
      - description: get number of per_page
        in: query
        name: per_page
        type: string
      - description: 'soft by cols exm: col1|asc,col2|desc'
        in: query
        name: sort
        type: string
      - description: 'search with column exm: col1|aaa,col2|bbb  by filter'
        in: query
        name: filter
        type: string
      - description: 'search by kw exm: aaa bbb ccc '
        in: query
        name: kw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Comments use VueTable
      tags:
      - comments
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"